TabletWeaving

Documentation for TabletWeaving.

TabletWeaving

Tablets

Tablet Structure

Tablet Threading

Tablet Stacking

Tablet

Tablet Rotation

Weaving

Tablets

A tablet or card is a square piece of card stock.

It has a front and a back.

It has a hole on each of it's four corners.

When looking at the front of the card, the holes are labeled A, B, C, and D clockwise.

The edges of the card are numbered, with the edge from corner A to corner B numbered 1. The edge between B and C numbered 2, and so on around the card.

For weaving, one warp thread passes through each hole.

The four warp threads for a tablet pass from the warp beam, through the tablet, to the cloth beam. They can pass through the tablet from BackToFront or FrontToBack.

Which thread passes through which hole, and how the tablet is threaded, can not change once the loom is warped.

It's easier to think about how the stitches are formed if the tablets are facing (either back or front depending on threading) the weaver. For compactness though the cards are stacked so that the weaver is facing their edges. When stacked, each tablet can be oriented FrontToTheRight or FrontToTheLeft.

Which direction a tablet is facing when it is stacked can be changed during weaving.

For each new row, each tablet must be rotated, and its stacking may be flipped. These operations affect the locations of the labeled holes (and their warp threads) relative to the loom.

Tablet Structure

We represent each edge of a tablet by a TabletEdge and each hole by a TabletHole:

TabletWeaving.TabletHole — Type

A TabletHole represents one of the corner holes in a tablet.

TabletWeaving.TabletEdge — Type

A TabletEdge` represents a single edge of a tablet.

The functions next, previousand opposite can beapplied to a TabletHole or a TabletEdge to get the next, previous, or opposite one respectively.

TabletWeaving.next_hole — Function

next_hole(::TabletEdge)

Return the hole following the specified edge.

TabletWeaving.previous_hole — Function

previous_hole(::TabletEdge)

Return the hole preceeding the specified hole.

Tablet Threading

How the warp threads pass through a tablet, the TabletThreading, can either be BackToFront or FrontToBack.

for BackToFront threading, warp threads pass from the warp beam through the tablet from back to front and then to the cloth beam.

For FrontToBack threading, Warp threads pass from the warp beam through the tablet from front to back and then to the cloth beam.

These threadings also have a concise textual representation. BackToFront can be represented by /or z. BackToFront can be represented by \\ or s.

TabletWeaving.BackToFront — Type

Warp threads pass from the warp beam through the tablet from back to front.

TabletWeaving.FrontToBack — Type

Warp threads pass from the warp beam through the tablet from front to back.

TabletWeaving.threading_for_char — Function

threading_for_char(::AbstractChar)

Given an 's' or a 'z', return an instance of the the corresponding TabletThreading.

There seems to be some confusion in the tablet weaving community about which is which. In this code base, 'z' corresponds to BackToFront threading.

TabletWeaving.threading_char — Function

threading_char(::TabletThreading)

Return the character ('s' or 'z' corresponding to the specified tablet threading. See threading_for_char.

Tablet Stacking

If the tablets are arranged with a flat side facing the weaver this would take too much space and also cause other inconveniences. Instead, tablets are arrahged so that their flat faces face each other. This forms a horizontal stack from one side of the loom to the other.

When stacked like this, a tablet is said to be stacked FrontToTheRight if it's front face faces towards the right side of the loom rom the weaver's perspective. A tablet facing the opposite direction is said to be threaded FrontToTheLeft.

For brevity in textual representations, FrontToTheRight is represented with a → right arrow and FrontToTheLeft with a ← left arrow.

TabletWeaving.TabletStacking — Type

TabletStacking is the abstract supertype for which way a tablet is facing when stacked.

TabletWeaving.FrontToTheRight — Type

In FrontToTheRight stacking, the tablet is stacked so that its front faces the weaver's right.

TabletWeaving.FrontToTheLeft — Type

In FrontToTheRight stacking, the tablet is stacked so that its front faces the weaver's left.

TabletWeaving.tablet_stacking_to_char — Function

tablet_stacking_to_char(::TabletStacking)

Return an arrow showing which direction the front of a tablet is currently facing.

Tablet

TabletWeaving.Tablet — Type

Tablet

Tablet describes the setup and current state of a single tablet weaving card.

TabletWeaving.warp_color — Function

warp_color(::Tablet, ::TabletHole)

Return the color of the warp thread that passes through the specified hole of the specified tablet.

TabletWeaving.top_edge — Function

top_edge(::Tablet)::TabletEdge

Return the TabletEdge of the top edge of the tablet. This edge is easier to see on the loom than the shed edge It is also unaffected by the tablet's stacking.

We use a Vector of Tablets to describe how the loom is set up for weaving. for convenience, we can add these vectors together for a wider pattern. We can also multiply them to repeat tablets.

Tablet Rotation

Prior to each new throw of the weft, the tablets are rotated to open a new shed.

In practice (the stacked arrangement), a card will be rotated forward or backward. Forward rotation moves the top corner of the card closest to the weaver away from the weaver towards the warp beam. Backward rotation moves the top corner furthest from the weaver toards the weaver.

Whether forward rotation turns the card in the ABCD or the DCBA direction depends on how the card is threaded and stacked.

TabletWeaving.RotationDirection — Type

RotationDirection

RotationDirection is the abstract supertype of all tablet rotations.

TabletWeaving.rotation — Function

rotation(::Tablet, ::RotationDirection)

Return the change in the Tablet's accumulated_rotation if the specified `AbstractRotation is applied.

TabletWeaving.ABCD — Type

The ABCD rotation causes the A corner of the tablet to move to the location in space previously occupied by the B corner.

TabletWeaving.DCBA — Type

The DCBA rotation causes the A corner of the tablet to move to the location in space previously occupied by the D corner.

TabletWeaving.Clockwise — Type

The Clockwise direction refers to how the tablet would move if its front or back face (depending on threading) were facing the weaver. Whether this results in ABCD or DCBA rotation depends on how the card is threaded.

TabletWeaving.CounterClockwise — Type

The CounterClockwise direction refers to how the tablet would move if its front or back face (depending on threading) were facing the weaver. Whether the front or the back of the tablet is facing the weaver depends on whether the card is BackToFrontorFrontToBack` threaded.

TabletWeaving.Forward — Type

The Forward rotation moves the top corner of the tablet closest to the weaver and the cloth beam to be the bottom corner closest to the weaver.

TabletWeaving.Backward — Type

The Backward rotation moves the top corner of the tablet closest to the weaver and the cloth beam to be the bottom corner closest to the weaver.

Weaving

To weave, the weaver first rotates the tablets according to some pattern. Rotating the tablets opens a new shed. The shuttle is then past through the new shed to finish weaving that row.

TabletWeaving.rotate! — Function

rotate!(::Tablet, ::RotationDirection)

Rotate the tablet by one position in the specified direction.

TabletWeaving.shot! — Function

shot!(::Tablet)

Apply the current rotation to the tablet and return the colors of the warp threads passing over the top and bottom of the fabric, and the crossing direction (as a forward or backslash character) when looking at that face of the fabric.

@docs simplerotationplan

### Going from an Image to a Pattern

We have an array of the "image" we want to weave.  How do we translate
that into a set of tablets, their warping, and their motions?

How do we execute that "plan" to produce a stitch image to see how the
pattern turned out.

For a two color pattern, we can warp each tablet with one color in
holes **A** and **C** and the other in holes **B** and **D**.
Whatever the previous stitch, the tablet can be rotated to either
color.  The slant of the stitch can't be controlled though.

@docs tabletsforimage

`tablets_for_image` does nothing about tablet threading, only colors.

The tablet weaving patterns I've been all seem to have one threading
on one side of the pattern and another threading on the other side,
with the possible exception of the borders having different threading
from the field.

We can introduce a function that sets one threading from the edge to
the middle and switches to the other threading for the other half.

@docs symetric_threading!

@docs TabletWeavingPattern

@index

@autodocs Modules = [TabletWeaving] ```